Update action metadata, improve .gitignore, and refactor project structure
- Renamed action from `msvc-dev-cmd` to `actions-msvc-dev-cmd` for clarity.
- Enhanced `action.yml` with updated description and author information.
- Expanded `.gitignore` to include additional files and directories.
- Updated test file paths in `.oxlintrc.json` for consistency.
- Removed unused files (`index.js`, `lib.js`, `version.js`, `vitest.config.js`) to streamline the project.
- Adjusted README for improved clarity and updated usage examples.
- Refactored build scripts to ensure compatibility with new structure.
