+++ .github/release.yml
@@ -3,20 +3,17 @@ changelog:
     labels:
       - ignore-for-release
       - github-actions
-    authors:
-      - octocat
-      - renovate[bot]
   categories:
-    - title: Breaking Changes 🛠
+    - title: Breaking Changes
       labels:
         - breaking-change
-    - title: Exciting New Features 🎉
+    - title: Features
       labels:
         - enhancement
         - feature
-    - title: Bug fixes 🐛
+    - title: Bug Fixes
       labels:
         - bug
-    - title: Other Changes 🔄
+    - title: Other Changes
       labels:
         - '*'
