+++ .github/workflows/ci.yml
@@ -11,7 +11,7 @@ env:
 on:
   workflow_dispatch:
   push:
-    branches: [master]
+    branches: [main]
   pull_request:
   schedule:
     - cron: '0 6 * * *'
