+++ .gitignore
@@ -1,4 +1,7 @@
-# actions requires a node_modules dir https://github.com/actions/toolkit/blob/master/docs/javascript-action.md#publish-a-releasesv1-action
-# but its recommended not to check these in https://github.com/actions/toolkit/blob/master/docs/action-versioning.md#recommendations
-node_modules
-coverage
+.DS_Store
+/.idea
+/node_modules
+/dist
+*.log
+coverage/
+.nyc_output/
