+++ .oxlintrc.json
@@ -7,7 +7,7 @@
   "ignorePatterns": ["dist/**", "node_modules/**", "coverage/**"],
   "overrides": [
     {
-      "files": ["**/*.test.js", "**/__tests__/**"],
+      "files": ["**/*.test.js", "**/tests/**"],
       "rules": {
         "require-yield": "off"
       }
