+++ README.md
@@ -1,138 +1,71 @@
-<a href="https://github.com/iShark5060/msvc-dev-cmd"><img alt="GitHub Actions status" src="https://github.com/iShark5060/msvc-dev-cmd/actions/workflows/main.yml/badge.svg"></a>
+# MSVC Dev Cmd
 
-# msvc-dev-cmd
+Configure the Microsoft Visual C++ Developer Command Prompt on Windows runners.
 
-[GitHub Action](https://github.com/features/actions) for configuring Developer Command Prompt for Microsoft Visual C++.
+[![CI](https://github.com/iShark5060/actions-msvc-dev-cmd/actions/workflows/ci.yml/badge.svg)](https://github.com/iShark5060/actions-msvc-dev-cmd/actions/workflows/ci.yml)
 
-This sets up the environment for compiling C/C++ code from command line.
+> **Fork notice:** Maintained fork of [ilammy/msvc-dev-cmd](https://github.com/ilammy/msvc-dev-cmd) by ilammy (MIT License).
+
+> **Always reference a published version tag** (e.g. `@v1`). The bundled action code (`dist/index.js`) is only committed to release tags, so referencing `@main` will not work.
 
 Supports Windows. Does nothing on Linux and macOS.
 
-> **Fork notice:** Maintained fork of [ilammy/msvc-dev-cmd](https://github.com/ilammy/msvc-dev-cmd). Original project by ilammy (MIT License).
-
-## Example usage
-
-Basic usage for default compilation settings is like this:
+## Usage
 
 ```yaml
 jobs:
-  test:
+  build:
+    runs-on: windows-latest
     steps:
       - uses: actions/checkout@v7
-      - uses: iShark5060/msvc-dev-cmd@v1
-      - name: Build something requiring CL.EXE
+      - uses: iShark5060/actions-msvc-dev-cmd@v1
+      - name: Build
         run: |
           cmake -G "NMake Makefiles" .
           nmake
-      # ...
 ```
 
-If you want something non-default,
-like using a specific version of Visual Studio,
-or cross-compling for a different target,
-you will need to configure those settings via inputs:
+Matrix builds for multiple architectures:
 
 ```yaml
 jobs:
-  test:
-    # Run a job for each of the specified target architectures:
+  build:
+    runs-on: windows-latest
     strategy:
       matrix:
-        arch:
-          - amd64
-          - amd64_x86
-          - amd64_arm64
+        arch: [amd64, amd64_x86, amd64_arm64]
     steps:
       - uses: actions/checkout@v7
-      - uses: iShark5060/msvc-dev-cmd@v1
+      - uses: iShark5060/actions-msvc-dev-cmd@v1
         with:
           arch: ${{ matrix.arch }}
-      - name: Build something requiring CL.EXE
+      - name: Build
         run: |
           cmake -G "NMake Makefiles" .
           nmake
-      # ...
 ```
 
 ## Inputs
 
-- `arch` – target architecture
-  - native compilation:
-    - `x64` (default) or its synonyms: `amd64`, `win64`, `x86_64`, `x86-64`
-    - `x86` or its synonyms: `win32`
-  - cross-compilation: `x86_amd64`, `x86_arm`, `x86_arm64`, `amd64_x86`, `amd64_arm`, `amd64_arm64`
-- `sdk` – Windows SDK to use
-  - do not specify to use the default SDK
-  - or specify full Windows 10 SDK number (e.g, `10.0.10240.0`)
-  - or write `8.1` to use Windows 8.1 SDK
-- `toolset` – select VC++ compiler toolset version
-  - do not specify to use the default toolset
-  - `14.0` for VC++ 2015 Compiler Toolset
-  - `14.XX` for the latest 14.XX toolset installed (e.g, `14.11`)
-  - `14.XX.YYYYY` for a specific full version number (e.g, `14.11.25503`)
-- `uwp` – set `true` to build for Universal Windows Platform (i.e., for Windows Store)
-- `spectre` – set `true` to use Visual Studio libraries with [Spectre](https://meltdownattack.com) mitigations
-- `vsversion` – the Visual Studio version to use. This can be the version number (e.g. 16.0 for 2019) or the year (e.g. "2019"); omit this input to select the latest version of Visual Studio. On [GitHub-hosted runners](https://docs.github.com/actions/using-github-hosted-runners/about-github-hosted-runners/about-github-hosted-runners), this input is not required because there is only one Visual Studio in the environment.
+| Input       | Default | Description                                                                  |
+| ----------- | ------- | ---------------------------------------------------------------------------- |
+| `arch`      | `x64`   | Target architecture (`x64`, `x86`, cross-compile variants like `amd64_x86`). |
+| `sdk`       | —       | Windows SDK version (e.g. `10.0.10240.0` or `8.1`).                          |
+| `toolset`   | —       | VC++ compiler toolset (e.g. `14.0`, `14.11`).                                |
+| `uwp`       | —       | Set `true` to build for Universal Windows Platform.                          |
+| `spectre`   | —       | Set `true` to use Spectre-mitigated Visual Studio libraries.                 |
+| `vsversion` | latest  | Visual Studio version number (e.g. `16.0`) or year (e.g. `2019`).            |
 
 ## Caveats
 
-### Name conflicts with `shell: bash`
+### `shell: bash` path conflicts
 
-Using `shell: bash` in Actions may shadow some of the paths added by MSVC.
-In particular, `link.exe` (Microsoft C linker) is prone to be shadowed by `/usr/bin/link` (GNU filesystem link tool).
-
-Unfortunately, this happens because GitHub Actions unconditionally _prepend_ GNU paths when `shell: bash` is used,
-on top of any paths set by `msvc-dev-cmd`, every time at the start of each new step.
-Hence, there aren't many non-destructive options here.
-
-If you experience compilation errors where `link` complains about unreasonable command-line arguments,
-“extra operand _something-something_” – that's probably it.
-Recommended workaround is to remove `/usr/bin/link` if that interferes with your builds.
-If this is not acceptable, please file an issue, then we'll figure out something better.
+GitHub Actions prepends GNU paths when `shell: bash` is used, which can shadow MSVC tools such as `link.exe`. If you see linker errors mentioning "extra operand", that is likely the cause.
 
 ### Reconfiguration
 
-You can invoke `iShark5060/msvc-dev-cmd` multiple times during your jobs with different inputs
-to reconfigure the environment for building with different settings
-(e.g., to target multiple architectures).
-
-```yaml
-jobs:
-  release:
-    steps:
-      # ...
-      - name: Configure build for amd64
-        uses: iShark5060/msvc-dev-cmd@v1
-        with:
-          arch: amd64
-
-      - run: build # (for amd64)
-
-      - name: Configure build for x86
-        uses: iShark5060/msvc-dev-cmd@v1
-        with:
-          arch: amd64_x86
-
-      - run: build # (for x86)
-
-      - name: Configure build for ARM64
-        uses: iShark5060/msvc-dev-cmd@v1
-        with:
-          arch: amd64_arm64
-
-      - run: build # (for ARM64)
-
-
-      # ...
-```
-
-This mostly works but it's not really recommended
-since Developer Command Prompt was not meant for recursive reconfiguration.
-That said, if it does not work for you, please file an issue.
-
-Consider using [`strategy.matrix`](https://docs.github.com/en/actions/reference/workflow-syntax-for-github-actions#jobsjob_idstrategymatrix)
-to execute different build configuration in parallel, independent environments.
+You can invoke the action multiple times in one job with different inputs, but using a `strategy.matrix` for parallel builds is preferred.
 
 ## License
 
-MIT, see [LICENSE](LICENSE).
+MIT. See [LICENSE](LICENSE).
