+++ action.yml
@@ -1,5 +1,7 @@
-name: Enable Developer Command Prompt
-description: Configure Microsoft Visual C++ command line
+name: MSVC Dev Cmd
+description: Configure Microsoft Visual C++ command line on Windows runners
+author: iShark5060
+
 inputs:
   arch:
     description: Target architecture
@@ -7,16 +9,18 @@ inputs:
   sdk:
     description: Windows SDK number to build for
   spectre:
-    description: Enable Specre mitigations
+    description: Enable Spectre mitigations
   toolset:
     description: VC++ compiler toolset version
   uwp:
     description: Build for Universal Windows Platform
   vsversion:
     description: The Visual Studio version to use. This can be the version number (e.g. 16.0 for 2019) or the year (e.g. "2019").
+
 runs:
-  using: 'node24'
+  using: node24
   main: dist/index.js
+
 branding:
   icon: terminal
   color: purple
