+++ package.json
@@ -1,27 +1,12 @@
 {
-  "name": "msvc-dev-cmd",
+  "name": "actions-msvc-dev-cmd",
   "version": "1.13.0-dev",
   "private": true,
   "description": "GitHub Action to setup Developer Command Prompt for Microsoft Visual C++",
-  "keywords": [
-    "Actions",
-    "Developer Command Prompt",
-    "GitHub",
-    "Microsoft Visual C++",
-    "Microsoft Visual Studio",
-    "Visual C++",
-    "Visual Studio",
-    "vcvarsall.bat"
-  ],
-  "homepage": "https://github.com/iShark5060/msvc-dev-cmd#readme",
-  "bugs": {
-    "url": "https://github.com/iShark5060/msvc-dev-cmd/issues"
-  },
   "license": "MIT",
-  "author": "ilammy <me@ilammy.net>",
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/iShark5060/msvc-dev-cmd.git"
+    "url": "git+https://github.com/iShark5060/actions-msvc-dev-cmd.git"
   },
   "type": "module",
   "main": "dist/index.js",
@@ -33,8 +18,8 @@
     "check-format": "oxfmt --check .",
     "test": "vitest run",
     "test:coverage": "vitest run --coverage",
-    "build": "esbuild index.js --bundle --platform=node --format=cjs --target=node24 --outfile=dist/index.js --minify",
-    "build-debug": "esbuild index.js --bundle --platform=node --format=cjs --target=node24 --outfile=dist/index.js --sourcemap --keep-names",
+    "build": "esbuild src/index.js --bundle --platform=node --format=cjs --target=node24 --outfile=dist/index.js --minify",
+    "build-debug": "esbuild src/index.js --bundle --platform=node --format=cjs --target=node24 --outfile=dist/index.js --sourcemap --keep-names",
     "deps": "pnpm dlx npm-check-updates --target newest -u --reject @types/node && pnpm dlx npm-check-updates @types/node --target semver -u",
     "deps:update": "pnpm update -r --latest && pnpm update @types/node"
   },
