+++ vitest.config.ts
@@ -3,9 +3,9 @@ import { defineConfig } from 'vitest/config';
 export default defineConfig({
   test: {
     environment: 'node',
-    include: ['__tests__/**/*.test.js'],
     coverage: {
       reporter: ['text', 'lcov'],
     },
+    include: ['tests/**/*.test.js'],
   },
 });
